test(worklets): move RetainingSerializable tests to their own file
